tools/hex2bdf.in: make it run on modern perl
Make it run on modern perl.
In modern perl, assignment of $[ is not permitted (for good reasons).

This script uses it to display the number of characters, which is done by
computing the last index of the array. Since the last index is off-by-one
compared to the length, they changed the first index instead of adding 1 to
the last index or using the array in scalar context to get its length.